Refactor Integrant system definition for better testability
For mock-able components, there are generic `:web-watchdog.system/...`
keywords that the concrete implementations derive from, notably:
 - ::system/app-state
 - ::system/download-fn
 - ::system/email-sender
 - ::system/app-state-observer

This will allow to supply test fakes for integration tests.

Using `integrant/resolve-key` where it makes sense - for the EmailSender
component, hiding its internal implementation details that are needed
only to stop the component.

Replacing composite keys with keyword hierarchies using `derive`. The
composite keys were a pain to work with (verbose, cumbersome to remember
that it's a vector of keywords instead of a single keyword which lead to
various annoying bugs).