Introduce new `:sort-elements-by-text` content extractor
This capability will be leveraged by inline-diff email notification
(coming later).

As a part of this change, it was necessary for `core/apply-content-extractor`
not to always convert the final result to a string because re-parsing
such a string into an org.jsoup.select.Elements object is quite cumbersome and
we are not guaranteed to reconstruct it correctly (e.g. a set of elements gets
wrapped into an "<html><head></head><body>..." structure. Instead we rely on
the content extractor chain to be well-formed and perform data type changes
automatically when needed.