Support emails with inline-diff change notification
Until now the email simply stated the old and new site content. When
combined with the `:sort-elements-by-text` content extractor, it makes
sense to also support showing a diff, with red/green highlighting the
changes.

This also required extending the state schema of a site to choose
between `"old-new"` and `"inline-diff"` notification formats. For that,
the old `:emails [<addresses>]` key was removed and replaced by
`:email-notification {:to [<addresses], :format <fmt>}`.