Fix email notification config in the Create site UI dialog
The 'Notification format' drop-down was being pre-populated with
'old-new' (a valid value), however the value wasn't being reflected
in the dialog data model. Submitting the dialog was then failing
server-side. I added into the dropdown a pre-selected empty value
with the '- select -' label.

For consistency, same handling was implemented for the Content
extractor type drow-down.