Notify web UI about state changes via server-sent events (SSE)
The previous mechanism of UI updates was based on periodic polling of the app
state, together with some spaghetti callbacks to force immediate state reloads
on particular user actions (e.g. add/edit a watched site, request site refresh).
SSE is a much nicer approach where the client is updated through a push model
instead of pull.

On the server, we subscribe to changes on the `app-state` atom and publish an
SSE event whenever the state changes (with sensible debouncing). The client
uses standard `EventSource` API to subscribe to the SSE events and when an
event is received, it reads the current state from the server.

The SSE ring handler is written as a synchronous handler (i.e. not using
asynchronous servlet API), which could normally hog threads. Therefore the
Jetty server was switched to use Java virtual threads which are the perfect
match for writing readable blocking I/O code that doesn't result in threadpool
depletion.

Sources:
- https://andersmurphy.com/2024/10/07/clojure-synchronous-server-sent-events-with-virtual-threads-and-channels.html
- https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/
- https://jetty.org/docs/jetty/12/programming-guide/arch/threads.html#thread-pool-virtual-threads