Support conditionally skipping email notifications
Added an optional site property `[:email-notification :condition]` that holds
a string with an S-expression condition. The condition needs to evaluate to true
for a notification to be sent.

Motivation: I have TfL Status checks for my commute time windows.
However, if the tube is disrupted at the end of the day and becomes
healthy the next morning, I get an unnecessary notification about the
change. Adding the following condition prevents that:

  (not (and check-succeeded?
            (content-matches? \"^$\")
            (< 3600 last-check-age-secs)))