Introduce an optional site property in `[:request :url]`
This URL serves as an override to be used when fetching the site
content, while using the top-level site `:url` property as a "display"
URL (e.g. in email notifications or web UI). This is useful if the site
content is consumed through an API end-point (json/xml/...) but the same
information is also presented to humans via a different URL.